requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
alembic==1.13.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
redis==5.0.8
celery==5.4.0
pynacl==1.5.0
jinja2==3.1.4
httpx==0.27.2
orjson==3.10.7
gunicorn==22.0.0
